Fix remaining compile errors: my missed imports + pre-existing bugs
My misses from the restructuring (Screens I forgot, only fixed their
ViewModels):
- RecipesScreen.kt, TemplatesScreen.kt, MealPlanScreen.kt: still had the
  old flat `import fr.scanneat.data.repository.*` (or no import at all
  for MealPlanScreen), which my earlier grep for specific class names
  didn't catch since these were bare wildcards. Pointed at the correct
  planning subpackage.
- WeightScreen.kt: needed `fr.scanneat.domain.engine.dashboard.WeightForecast`.
- OffMapper.kt (nutrition) needed `inferCategoryFromName` from the
  scoring subpackage — missed during the Step 3 cross-reference check.

Pre-existing bugs, unrelated to the restructuring (verified via git diff
against the original extraction commit — none of these lines were
touched before now):
- app/build.gradle.kts: added module-wide `-opt-in=
  androidx.compose.material3.ExperimentalMaterial3Api`. ~14 screens call
  TopAppBar/TopAppBarDefaults without @OptIn, which is a hard compile
  error (not a warning) without the opt-in — cheaper to declare once
  than annotate every screen.
- OpenFoodFactsApi.kt: OFF_FIELDS was `const val` but initialized with
  `.joinToString(...)`, which isn't a compile-time constant.
- ConsumptionRepository/CustomFoodRepository called dao.observeRange()/
  dao.deleteByName() — methods that never existed on ConsumptionDao/
  CustomFoodDao (confirmed against the original flat Daos.kt). Added
  both DAO query methods.
- DayNotesRepository.kt, ComparisonRepository.kt: missing
  `import kotlinx.coroutines.flow.first`.
- MealPlanRepository.kt: destructuring a 5-element List failed to
  resolve component5(); replaced with plain indexed access.
- DiaryViewModel.kt: missing `import fr.scanneat.domain.model.ConsumedNutrition`.
- HydrationScreen.kt/HydrationViewModel.kt: missing imports for the
  top-level HYD_GLASS_ML/HYD_DEFAULT_GOAL_ML constants; the ViewModel
  also incorrectly qualified them as `HydrationRepository.HYD_...`,
  which doesn't work since they're top-level, not companion members.
- DataScreen.kt: a `forEach { (labelVal, setter) -> val (lv, set) =
  labelVal; val (l, v) = lv }` destructured one nesting level too many,
  trying to destructure a String. Removed the extra level.
- TrackerScreen.kt: `fatFrac/carbFrac/protFrac` are Double;
  `.coerceAtLeast(0.01f)` passed a Float literal to a Double receiver.
  Changed literals to `0.01`.
- DiaryScreen.kt: `collectAsStateWithLifecycle(initial = ...)` — the
  actual parameter name is `initialValue`.
- ResultScreen.kt (previous commit missed this one): `pillar.score`
  (Double) divided by `pillar.max.toFloat()` promotes to Double, so
  `.coerceIn(0f, 1f)` mismatched; added `.toFloat()` on both operands.

+++ scan-eat-android/app/src/main/java/fr/scanneat/presentation/result/ResultScreen.kt
@@ -374,7 +374,7 @@ private fun PillarsSection(pillars: ScoreAudit.Pillars) {
 
 @Composable
 private fun PillarRow(pillar: PillarScore) {
-    val ratio = (pillar.score / pillar.max.toFloat()).coerceIn(0f, 1f)
+    val ratio = (pillar.score.toFloat() / pillar.max.toFloat()).coerceIn(0f, 1f)
     val color = when { ratio >= 0.7f -> FlagGreen; ratio >= 0.4f -> AmberWarning; else -> FlagRed }
     Column(modifier = Modifier.fillMaxWidth()) {
         Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
